pyproject.toml
[project]
name = "projet-final"
version = "0.1.0"
description = ""
authors = [
    {name = "aichaa28",email = "aicha.ettriki28@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit (>=1.41.1,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "matplotlib (>=3.10.0,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
ruff = "^0.9.4"
pre-commit = "^4.1.0"


[tool.poetry.group.docs.dependencies]
sphinx = "^8.1.3"
sphinx-rtd-theme = "^3.0.2"

